Fixing to get TFMX player working (which it does now, yah!)

+++ src/plugins/tfmx/src/tfmx.h
@@ -12,10 +12,10 @@
 
 /* arch-dependent stuff here */
 
-typedef unsigned long U32;
+typedef unsigned int U32;
 typedef unsigned short U16;
 typedef unsigned char U8;
-typedef long S32;
+typedef int S32;
 typedef short S16;
 typedef char S8;
